scan stability and dry run improvements
