+++ printers/image.go
@@ -3,7 +3,6 @@ package printers
 import (
 	"image"
 	"image/color"
-	"math"
 	"sort"
 
 	"github.com/disintegration/imaging"
@@ -13,13 +12,15 @@ import (
 
 const (
 	DefaultThreshold = 128 // Default threshold for dark pixels
+	DefaultGamma     = 0.0
 )
 
-var ditherFunctions = map[string]func(image.Image) image.Image{
+var ditherFunctions = map[string]func(image.Image, float64) image.Image{
 	"floyd-steinberg": dFloydSteinberg,
 	"atkinson":        dAtkinson,
 	"stucki":          dStucki,
 	"bayer":           dBayer,
+	"no-dither":       DitherThresholdFn(DefaultThreshold),
 }
 
 func AllDitherFunctions() []string {
@@ -53,240 +54,54 @@ func resize(img image.Image, targetWidth int) image.Image {
 }
 
 // ditherimg is the default dither function used in the rasteriser.
-func ditherimg(img image.Image) image.Image {
-	return dAtkinson(img) // default dithering function
+func ditherimg(img image.Image, gamma float64) image.Image {
+	return dAtkinson(img, gamma) // default dithering function
 }
 
-func dFloydSteinberg(img image.Image) image.Image {
-	dstImage := imaging.AdjustGamma(img, 1.5)
+func dFloydSteinberg(img image.Image, gamma float64) image.Image {
+	const defaultGamma = 1.5
+	if gamma == 0.0 {
+		gamma = defaultGamma
+	}
+	dstImage := imaging.AdjustGamma(img, gamma)
 	dithered := image.NewPaletted(img.Bounds(), []color.Color{color.Black, color.White})
 	// increase brightness of the image to make it more suitable for dithering
 	draw.FloydSteinberg.Draw(dithered, dithered.Bounds(), dstImage, image.Point{})
 	return dithered
 }
 
-func dAtkinson(img image.Image) image.Image {
+func dAtkinson(img image.Image, gamma float64) image.Image {
+	const defaultGamma = 3.0
+	if gamma == 0.0 {
+		gamma = defaultGamma
+	}
 	dithered := image.NewRGBA(img.Bounds())
 	d := dither.NewDitherer([]color.Color{color.Black, color.White})
 	d.Matrix = dither.Atkinson
-	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, 3.0), image.Point{})
+	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, gamma), image.Point{})
 	return dithered
 }
 
-func dStucki(img image.Image) image.Image {
+func dStucki(img image.Image, gamma float64) image.Image {
+	const defaultGamma = 3.5
+	if gamma == 0.0 {
+		gamma = defaultGamma
+	}
 	dithered := image.NewRGBA(img.Bounds())
 	d := dither.NewDitherer([]color.Color{color.Black, color.White})
 	d.Matrix = dither.Stucki
-	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, 3.5), image.Point{})
+	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, gamma), image.Point{})
 	return dithered
 }
 
-func dBayer(img image.Image) image.Image {
+func dBayer(img image.Image, gamma float64) image.Image {
+	const defaultGamma = 3.5
+	if gamma == 0.0 {
+		gamma = defaultGamma
+	}
 	dithered := image.NewRGBA(img.Bounds())
 	d := dither.NewDitherer([]color.Color{color.Black, color.White})
 	d.Mapper = dither.Bayer(8, 8, 1.0) // 8x8 Bayer matrix
-	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, 3.5), image.Point{})
+	d.Draw(dithered, dithered.Bounds(), imaging.AdjustGamma(img, gamma), image.Point{})
 	return dithered
-}
-
-func resizeAndDither(img image.Image, targetWidth int, ditherFn func(image.Image) image.Image) image.Image {
-	return ditherFn(resize(img, targetWidth))
-}
-
-type Raster struct {
-	Width          int
-	Dpi            int
-	LinesPerPacket int
-	PrefixFunc     func(packetIndex int) []byte      // returns 55 m n
-	Terminator     byte                              // 00
-	DitherFunc     func(img image.Image) image.Image // optional dither function
-	Threshold      uint8                             // threshold for dark pixels, default is 128
-}
-
-type Rasteriser interface {
-	Rasterise(src image.Image) [][]byte                 // returns a slice of byte slices, each representing a packet
-	DPI() int                                           // returns the DPI of the rasteriser
-	LineWidth() int                                     // returns the line width in pixels
-	SetDitherFunc(fn func(img image.Image) image.Image) // sets the dither function
-}
-
-func DitherThresholdFn(threshold uint8) func(img image.Image) image.Image {
-	return func(img image.Image) image.Image {
-		if threshold == 0 {
-			threshold = DefaultThreshold // default threshold for dark pixels
-		}
-		trg := image.NewPaletted(img.Bounds(), []color.Color{color.Black, color.White})
-		for x := img.Bounds().Min.X; x < img.Bounds().Max.X; x++ {
-			for y := img.Bounds().Min.Y; y < img.Bounds().Max.Y; y++ {
-				if pixelBit(img, x, y, threshold) {
-					trg.SetColorIndex(x, y, 0) // black
-				} else {
-					trg.SetColorIndex(x, y, 1) // white
-				}
-			}
-		}
-		return trg
-	}
-}
-
-func (r *Raster) DPI() int {
-	return r.Dpi
-}
-
-func (r *Raster) LineWidth() int {
-	return r.Width
-}
-
-func (r *Raster) SetDitherFunc(fn func(img image.Image) image.Image) {
-	if fn == nil {
-		r.DitherFunc = dFloydSteinberg // reset to default if nil
-	} else {
-		r.DitherFunc = fn
-	}
-}
-
-func (r *Raster) Rasterise(src image.Image) [][]byte {
-	var (
-		msgPrefixSz     = len(r.PrefixFunc(0)) // 55 m n
-		msgTerminatorSz = 1                    // 00
-
-		lineWidthPixels = r.Width
-		lineWidthBytes  = lineWidthPixels / 8
-		linesPerMsg     = r.LinesPerPacket
-
-		msgDataSz    = lineWidthBytes * linesPerMsg
-		msgPayloadSz = msgPrefixSz + msgDataSz + msgTerminatorSz // 55 m n + data + 00
-	)
-
-	dfn := ditherimg
-	if r.DitherFunc != nil {
-		dfn = r.DitherFunc
-	}
-
-	resized := resize(src, lineWidthPixels)
-	img := resized
-	if !isDocument(resized, 50, 200) {
-		// If the image is not a document, apply dithering
-		img = dfn(resized)
-	}
-	debugSaveImage(img, "rasterised.png")
-
-	bounds := img.Bounds()
-	width := bounds.Dx()
-	height := bounds.Dy()
-
-	if width > lineWidthPixels {
-		panic("image width exceeds 384px limit")
-	}
-
-	rasteriseLine := func(img image.Image, y int) []byte {
-		lineBytes := make([]byte, lineWidthBytes)
-		for x := range lineWidthPixels {
-			bit := pixelBit(img, bounds.Min.X+x, bounds.Min.Y+y, r.Threshold)
-			if bit {
-				lineBytes[x/8] |= (1 << (7 - (x % 8)))
-			}
-		}
-		return lineBytes
-	}
-
-	// Pad height to even number for 2-line packets
-	if height%2 != 0 {
-		height++
-	} else {
-		height += 2 // ensure we have at least 2 lines for the last packet
-	}
-
-	numPackets := height / linesPerMsg
-	packets := make([][]byte, 0, numPackets)
-
-	for packetIndex := range numPackets {
-		y0 := packetIndex * 2
-		y1 := y0 + 1
-
-		row := make([]byte, 0, msgPayloadSz)
-
-		row = append(row, r.PrefixFunc(packetIndex)...)
-
-		// First line (y0)
-		lineBytes := rasteriseLine(img, y0)
-		row = append(row, lineBytes...)
-
-		// Second line (y1)
-		lineBytes = rasteriseLine(img, y1)
-		row = append(row, lineBytes...)
-
-		row = append(row, r.Terminator) // terminating byte
-
-		packets = append(packets, row)
-	}
-
-	return packets
-}
-
-func pixelBit(img image.Image, x, y int, threshold uint8) bool {
-	if threshold == 0 {
-		threshold = DefaultThreshold // default threshold for dark pixels
-	}
-	if y >= img.Bounds().Dy() {
-		return false // padded line
-	}
-	if x >= img.Bounds().Dx() {
-		return false // image narrower than 384px
-	}
-
-	c := img.At(x, y)
-	gray := colorToGray(c)
-	return gray < threshold // dark pixels are "on"
-}
-
-func colorToGray(c color.Color) uint8 {
-	if gray, ok := c.(color.Gray); ok {
-		return gray.Y
-	}
-	r, g, b, _ := c.RGBA()
-	gray := (299*r + 587*g + 114*b) / 1000
-	return uint8(gray >> 8)
-}
-
-func isDocument(img image.Image, darkThreshold, lightThreshold uint8) bool {
-	if img == nil {
-		return false
-	}
-	if darkThreshold == 0 {
-		darkThreshold = 50
-	}
-	if lightThreshold == 0 {
-		lightThreshold = 200
-	}
-	bounds := img.Bounds()
-	dst := image.NewGray(img.Bounds())
-	draw.Draw(dst, bounds, img, image.Point{}, draw.Src)
-	// create histogram of pixel brightness
-	histogram := make([]int, math.MaxUint8+1)
-	for y := bounds.Min.Y; y < bounds.Max.Y; y++ {
-		for x := bounds.Min.X; x < bounds.Max.X; x++ {
-			c := dst.At(x, y).(color.Gray)
-			histogram[c.Y]++
-		}
-	}
-	// sum all dark pixels in the range [0, darkThreshold)
-	var (
-		darkPixelCount  float64
-		lightPixelCount float64
-		totalPixelCount float64
-	)
-	for i, count := range histogram {
-		totalPixelCount += float64(count)
-		if i < int(darkThreshold) {
-			darkPixelCount += float64(count)
-		} else if i >= int(lightThreshold) {
-			lightPixelCount += float64(count)
-		}
-	}
-	if totalPixelCount == 0 {
-		return false // no pixels to analyze
-	}
-
-	return (darkPixelCount+lightPixelCount)/totalPixelCount > 0.85
 }
